+++ test/libsolidity/syntaxTests/inheritance/override/public_var_override_mapping_to_dynamic_struct.sol
@@ -1,0 +1,12 @@
+pragma experimental ABIEncoderV2;
+
+abstract contract C {
+	struct S {
+	uint256 x;
+	string y;
+	}
+	function f(address x) external virtual returns (uint256, string memory);
+}
+contract D is C {
+	mapping(address => S) public override f;
+}